refactor(ui): split background task rail rendering
