+++ ui/src/pages/chat/components/chat-background-tasks-status.ts
@@ -12,7 +12,8 @@ import {
   taskTitle,
 } from "../../../lib/tasks/data.ts";
 import type { TaskSummary } from "../../../lib/tasks/task-summary.ts";
-import { STATUS_TONES, type BackgroundTasksProps } from "./chat-background-tasks.ts";
+import { STATUS_TONES } from "./chat-background-tasks-shared.ts";
+import type { BackgroundTasksProps } from "./chat-background-tasks.types.ts";
 
 type BackgroundTasksStatus = { count: number; startedMs: number | null };
